+++ extra/constants.js
@@ -1,0 +1,9 @@
+export const permissions = 
+{
+    'getUsers': {
+    all: ['head-trainer'],
+    read : ['trainee', 'trainer'],
+    write : ['trainer'],
+    delete: [],
+    }
+};
