+++ extra/patterns/index.js
@@ -1,0 +1,4 @@
+import diamond from './diamond';
+import equilateral from './equilateral';
+
+export {equilateral, diamond};
