+++ extra/utils/helpers.js
@@ -1,0 +1,6 @@
+export const validateEmail=(email)=>{
+        
+    var validRegex = /^[a-zA-Z0-9.^]+@successive.tech/;
+
+    return validRegex.test(email) 
+}
